ingestion/requirements.txt
psycopg2-binary==2.9.3
json5==0.9.10 